Add BlockIndex to ContentIndex and rewrite parseTokens tests with recording DiffProducer
- Add ContentIndex.BlockIndex(blockIndex) for block-level producer calls
- Replace parseTokens tests with a recording DiffProducer that collects
  List<Pair<ContentIndex, Change<*>>>, covering block(), itemList(), and textSegment()
- Tests now verify structural producer calls (block insert/delete, itemList insert)
  in addition to text segment changes
- Split EditLetterWordTokenizerTest from EditLetterWordDiffTest (completed in prior commit)
